fix: modify files in launchpad homepage
