+++ src/app/page.tsx
@@ -1,12 +1,12 @@
 import Link from 'next/link';
+import Footer from '@/components/footer';
+import CopyButton from '@/components/copy-button';
 
-import CopyButton from '../components/copybutton';
 
 export default function Home() {
-  const currentYear = new Date().getFullYear();
   return (
     <>
-      <section className="w-full py-12  lg:py-14 xl:py-24 2xl:py-28">
+      <section className="min-h-[94vh] w-full py-12 lg:py-14 xl:py-24 2xl:py-28">
         <div className="container mx-auto space-y-8 px-4 md:px-6">
           <div className="flex flex-col items-center space-y-4 text-center">
             <div className="space-y-4">
@@ -37,392 +37,10 @@ export default function Home() {
           </div>
         </div>
       </section>
-      <footer className="items-center justify-evenly bg-gray-300 py-4 text-center text-black lg:flex">
-        <Link href="https://www.creowis.com/" target="_blank">
-          &copy; {currentYear} An Open Source initiative from CreoWis
-          Technologies.
-        </Link>
-        <div className="mt-4 flex justify-center gap-4 lg:mt-0 lg:gap-8">
-          {SOCIAL_ICONS.map((item, id) => (
-            <Link key={id} href={item.link} target="_blank">
-              <div className=" hover:scale-105">{item.icon}</div>
-            </Link>
-          ))}
-        </div>
-      </footer>
+      <Footer />
     </>
   );
 }
-
-const ICONS = {
-  next_js: (
-    <svg
-      className="h-6 w-6"
-      fill="none"
-      height="15"
-      viewBox="0 0 15 15"
-      width="15"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="m4.5 4.5.40534-.29275c-.12698-.17581-.35282-.24972-.55917-.183-.20636.06673-.34617.25888-.34617.47575zm3 9.5c-3.58985 0-6.5-2.9101-6.5-6.5h-1c0 4.1421 3.35786 7.5 7.5 7.5zm6.5-6.5c0 3.5899-2.9101 6.5-6.5 6.5v1c4.1421 0 7.5-3.3579 7.5-7.5zm-6.5-6.5c3.5899 0 6.5 2.91015 6.5 6.5h1c0-4.14214-3.3579-7.5-7.5-7.5zm0-1c-4.14214 0-7.5 3.35786-7.5 7.5h1c0-3.58985 2.91015-6.5 6.5-6.5zm-2.5 12v-7.5h-1v7.5zm-.90534-7.20725 6.50004 8.99995.8106-.5854-6.49996-9.00005zm5.90534-.79275v6h1v-6z"
-        fill="#000"
-      />
-    </svg>
-  ),
-
-  typescript: (
-    <svg
-      className="h-6 w-6"
-      height="256"
-      preserveAspectRatio="xMidYMid"
-      viewBox="0 0 256 256"
-      width="256"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="m0 128v-128h128 128v128 128h-128-128z"
-        fill="#007acc"
-        transform="matrix(1 0 0 -1 0 256)"
-      />
-      <g fill="#fff">
-        <path
-          d="m146.658132 223.436863.081269-10.483809h-16.660317-16.660317v-47.339683-47.339682h-11.784127-11.7841274v47.339682 47.339683h-16.6603175-16.6603175v10.280635c0 5.688888.1219048 10.443174.2844445 10.565079.1219048.16254 20.3987301.243809 44.9828569.203175l44.739048-.121905z"
-          transform="matrix(-1 0 0 -1 203.26928 352.285986)"
-        />
-        <path
-          d="m206.566631 234.272145c6.501588-1.625397 11.459048-4.510477 16.010159-9.224127 2.356826-2.519365 5.851429-7.111111 6.135873-8.208254.08127-.32508-11.052698-7.801905-17.798095-11.987302-.24381-.16254-1.219048.893968-2.316191 2.519365-3.291428 4.794921-6.745396 6.867302-12.027936 7.233016-7.76127.528254-12.759365-3.535238-12.71873-10.32127 0-1.991111.284444-3.169524 1.097143-4.79492 1.706666-3.535239 4.87619-5.648254 14.831746-9.955556 18.326349-7.883175 26.168889-13.084444 31.045079-20.48 5.445079-8.248889 6.664127-21.414603 2.966349-31.207619-4.063492-10.646349-14.140952-17.879365-28.322539-20.276825-4.388572-.772064-14.791112-.650159-19.504762.203174-10.280635 1.828572-20.033016 6.907937-26.046984 13.572064-2.356826 2.600635-6.948572 9.386666-6.664127 9.874285.121904.16254 1.178412.812699 2.356825 1.503492 1.137778.650159 5.445079 3.128889 9.508571 5.485715l7.354921 4.266666 1.544127-2.275555c2.153651-3.291429 6.867302-7.801905 9.711746-9.305397 8.167619-4.307302 19.382857-3.697778 24.909206 1.259683 2.356826 2.15365 3.332064 4.388571 3.332064 7.68 0 2.966349-.365714 4.266666-1.909841 6.501587-1.991112 2.844444-6.054604 5.241905-17.594921 10.24-13.206349 5.688889-18.895238 9.224127-24.096508 14.831746-3.006984 3.250794-5.851429 8.452063-7.029841 12.8-.975238 3.616508-1.219048 12.678095-.446984 16.335238 2.722539 12.759365 12.353015 21.658413 26.250158 24.299683 4.510477.853333 14.994286.528253 19.423492-.568889z"
-          transform="matrix(1 0 0 -1 0 352.38048)"
-        />
-      </g>
-    </svg>
-  ),
-
-  tailwind: (
-    <svg
-      className="h-6 w-6"
-      viewBox="0 0 32 32"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="m9 13.7q1.4-5.6 7-5.6c5.6 0 6.3 4.2 9.1 4.9q2.8.7 4.9-2.1-1.4 5.6-7 5.6c-5.6 0-6.3-4.2-9.1-4.9q-2.8-.7-4.9 2.1zm-7 8.4q1.4-5.6 7-5.6c5.6 0 6.3 4.2 9.1 4.9q2.8.7 4.9-2.1-1.4 5.6-7 5.6c-5.6 0-6.3-4.2-9.1-4.9q-2.8-.7-4.9 2.1z"
-        fill="#44a8b3"
-      />
-    </svg>
-  ),
-
-  eslint: (
-    <svg
-      className="h-6 w-6"
-      height="225"
-      preserveAspectRatio="xMidYMid"
-      viewBox="0 0 256 225"
-      width="256"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="m77.9645944 80.5683518 48.5703096-28.0418525c1.215461-.7019991 2.713728-.7019991 3.929189 0l48.571313 28.0418525c1.214458.701999 1.964594 1.9986915 1.964594 3.4026895v56.0837047c0 1.402996-.750136 2.700691-1.964594 3.403693l-48.571313 28.041852c-1.215461.701999-2.713728.701999-3.929189 0l-48.5703096-28.041852c-1.2154612-.701999-1.9645944-1.999695-1.9645944-3.403693v-56.0837047c.0010029-1.403998.7501361-2.7006905 1.9645944-3.4026895"
-        fill="#8080f2"
-      />
-      <path
-        d="m254.4167 107.416596-58.093559-101.06679208c-2.11023-3.65427451-6.008432-6.34980392-10.227888-6.34980392h-116.1891259c-4.2204594 0-8.1196659 2.69552941-10.2298956 6.34980392l-58.09355923 100.84392108c-2.11022969 3.655279-2.11022969 8.268299 0 11.923577l58.09355923 100.238557c2.1102297 3.654274 6.0094362 5.522572 10.2298956 5.522572h116.1881219c4.219456 0 8.118662-1.812078 10.227888-5.467356l58.093559-100.402197c2.112238-3.65327 2.112238-7.938007.001004-11.592282zm-48.104603 48.599843c0 1.485804-.894488 2.860173-2.182512 3.604079l-73.998587 42.693772c-1.287019.742902-2.899307.742902-4.186326 0l-74.0558105-42.693772c-1.2870193-.743906-2.1875312-2.117271-2.1875312-3.604079v-85.3885488c0-1.4868078.887461-2.8601726 2.1754842-3.6040784l73.9955755-42.6937726c1.287019-.7439059 2.897299-.7439059 4.184318 0l74.06083 42.6937726c1.288024.7439058 2.194559 2.1172706 2.194559 3.6040784z"
-        fill="#4b32c3"
-      />
-    </svg>
-  ),
-
-  prettier: (
-    <svg
-      className="h-6 w-6"
-      viewBox="0 0 32 32"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="m21.714 8.571h1.143a.571.571 0 0 1 .571.571.571.571 0 0 1 -.571.571h-1.143a.571.571 0 0 1 -.571-.571.571.571 0 0 1 .571-.571z"
-        fill="#56b3b4"
-      />
-      <path
-        d="m4.571 26.857h5.714a.571.571 0 0 1 .571.571.571.571 0 0 1 -.571.571h-5.714a.571.571 0 0 1 -.571-.57.571.571 0 0 1 .571-.572z"
-        fill="#ea5e5e"
-      />
-      <path
-        d="m18.286 17.714h3.429a.571.571 0 0 1 .571.571.571.571 0 0 1 -.571.571h-3.429a.571.571 0 0 1 -.571-.571.571.571 0 0 1 .571-.571z"
-        fill="#bf85bf"
-      />
-      <path
-        d="m11.429 17.714h4.571a.571.571 0 0 1 .571.571.571.571 0 0 1 -.571.571h-4.571a.571.571 0 0 1 -.571-.571.571.571 0 0 1 .571-.571z"
-        fill="#ea5e5e"
-      />
-      <path
-        d="m4.571 17.714h4.572a.571.571 0 0 1 .571.571.571.571 0 0 1 -.571.571h-4.572a.571.571 0 0 1 -.571-.57.571.571 0 0 1 .571-.572z"
-        fill="#56b3b4"
-      />
-      <path
-        d="m4.571 22.286h5.714a.571.571 0 0 1 .571.571.571.571 0 0 1 -.571.571h-5.714a.571.571 0 0 1 -.571-.571.571.571 0 0 1 .571-.571z"
-        fill="#bf85bf"
-      />
-      <path
-        d="m4.571 13.143h5.714a.571.571 0 0 1 .571.571.571.571 0 0 1 -.571.571h-5.714a.571.571 0 0 1 -.571-.571.571.571 0 0 1 .571-.571z"
-        fill="#bf85bf"
-      />
-      <path
-        d="m10.286 6.286h11.428a.571.571 0 0 1 .571.571.571.571 0 0 1 -.571.571h-11.428a.571.571 0 0 1 -.571-.571.571.571 0 0 1 .571-.571z"
-        fill="#f7ba3e"
-      />
-      <path
-        d="m4.571 6.286h3.429a.571.571 0 0 1 .571.571.571.571 0 0 1 -.571.572h-3.429a.571.571 0 0 1 -.571-.572.571.571 0 0 1 .571-.571z"
-        fill="#ea5e5e"
-      />
-      <path
-        d="m9.143 24.571h1.143a.571.571 0 0 1 .571.571.571.571 0 0 1 -.571.571h-1.143a.571.571 0 0 1 -.571-.571.571.571 0 0 1 .571-.571z"
-        fill="#f7ba3e"
-      />
-      <path
-        d="m9.143 10.857h1.143a.571.571 0 0 1 .571.571.571.571 0 0 1 -.571.571h-1.143a.571.571 0 0 1 -.571-.571.571.571 0 0 1 .571-.571z"
-        fill="#56b3b4"
-      />
-      <path
-        d="m4.571 24.571h2.286a.571.571 0 0 1 .571.571.571.571 0 0 1 -.571.571h-2.286a.571.571 0 0 1 -.571-.57.571.571 0 0 1 .571-.572z"
-        fill="#56b3b4"
-      />
-      <path
-        d="m4.571 10.857h2.286a.571.571 0 0 1 .571.571.571.571 0 0 1 -.571.572h-2.286a.571.571 0 0 1 -.571-.571.571.571 0 0 1 .571-.572z"
-        fill="#f7ba3e"
-      />
-      <path
-        d="m19.429 24.571h1.143a.571.571 0 0 1 .571.571.571.571 0 0 1 -.571.571h-1.143a.571.571 0 0 1 -.571-.571.571.571 0 0 1 .571-.571z"
-        fill="#d0d4d8"
-        opacity=".5"
-      />
-      <path
-        d="m12.571 24.571h4.571a.571.571 0 0 1 .571.571.571.571 0 0 1 -.571.571h-4.571a.571.571 0 0 1 -.571-.57.571.571 0 0 1 .571-.572z"
-        fill="#d0d4d8"
-        opacity=".5"
-      />
-      <path
-        d="m22.857 24.571h4.571a.571.571 0 0 1 .571.571.571.571 0 0 1 -.571.571h-4.571a.571.571 0 0 1 -.571-.571.571.571 0 0 1 .571-.571z"
-        fill="#d0d4d8"
-        opacity=".5"
-      />
-      <path
-        d="m13.714 15.429h9.143a.571.571 0 0 1 .571.571.571.571 0 0 1 -.571.571h-9.143a.571.571 0 0 1 -.571-.571.571.571 0 0 1 .571-.571z"
-        fill="#56b3b4"
-      />
-      <path
-        d="m8 15.429h3.429a.571.571 0 0 1 .571.571.571.571 0 0 1 -.571.571h-3.429a.571.571 0 0 1 -.571-.571.571.571 0 0 1 .571-.571z"
-        fill="#f7ba3e"
-      />
-      <path
-        d="m4.571 15.429h1.143a.571.571 0 0 1 .572.571.571.571 0 0 1 -.571.571h-1.144a.571.571 0 0 1 -.571-.571.571.571 0 0 1 .571-.571z"
-        fill="#ea5e5e"
-      />
-      <path
-        d="m14.857 8.571h4.571a.571.571 0 0 1 .572.572.571.571 0 0 1 -.571.571h-4.572a.571.571 0 0 1 -.571-.571.571.571 0 0 1 .571-.572z"
-        fill="#bf85bf"
-      />
-      <path
-        d="m4.571 8.571h8a.571.571 0 0 1 .571.571.571.571 0 0 1 -.571.571h-8a.571.571 0 0 1 -.571-.57.571.571 0 0 1 .571-.572z"
-        fill="#56b3b4"
-      />
-      <path
-        d="m8 20h10.286a.571.571 0 0 1 .571.571.571.571 0 0 1 -.571.571h-10.286a.571.571 0 0 1 -.571-.571.571.571 0 0 1 .571-.571z"
-        fill="#f7ba3e"
-      />
-      <path
-        d="m4.571 20h1.143a.571.571 0 0 1 .571.571.571.571 0 0 1 -.571.571h-1.143a.571.571 0 0 1 -.571-.571.571.571 0 0 1 .571-.571z"
-        fill="#bf85bf"
-      />
-      <path
-        d="m18.286 10.857h5.714a.571.571 0 0 1 .571.571.571.571 0 0 1 -.571.572h-5.714a.571.571 0 0 1 -.571-.571.571.571 0 0 1 .571-.572z"
-        fill="#ea5e5e"
-      />
-      <path
-        d="m18.286 13.143h5.714a.571.571 0 0 1 .571.571.571.571 0 0 1 -.571.571h-5.714a.571.571 0 0 1 -.571-.571.571.571 0 0 1 .571-.571z"
-        fill="#f7ba3e"
-      />
-      <path
-        d="m4.571 4h13.715a.571.571 0 0 1 .571.571.571.571 0 0 1 -.571.571h-13.715a.571.571 0 0 1 -.571-.571.571.571 0 0 1 .571-.571z"
-        fill="#56b3b4"
-      />
-      <g fill="#d0d4d8">
-        <path
-          d="m20.571 4h6.857a.571.571 0 0 1 .572.571.571.571 0 0 1 -.571.571h-6.858a.571.571 0 0 1 -.571-.571.571.571 0 0 1 .571-.571z"
-          opacity=".5"
-        />
-        <path
-          d="m20.571 20h2.286a.571.571 0 0 1 .571.571.571.571 0 0 1 -.571.571h-2.286a.571.571 0 0 1 -.571-.571.571.571 0 0 1 .571-.571z"
-          opacity=".5"
-        />
-        <path
-          d="m25.143 20h2.286a.571.571 0 0 1 .571.571.571.571 0 0 1 -.571.571h-2.286a.571.571 0 0 1 -.571-.571.571.571 0 0 1 .571-.571z"
-          opacity=".5"
-        />
-        <path
-          d="m24 17.714h3.429a.571.571 0 0 1 .571.571.571.571 0 0 1 -.571.571h-3.429a.571.571 0 0 1 -.571-.571.571.571 0 0 1 .571-.571z"
-          opacity=".5"
-        />
-        <path
-          d="m24 6.286h3.429a.571.571 0 0 1 .571.571.571.571 0 0 1 -.571.571h-3.429a.571.571 0 0 1 -.571-.571.571.571 0 0 1 .571-.571z"
-          opacity=".5"
-        />
-        <path
-          d="m25.143 15.429h2.286a.571.571 0 0 1 .571.571.571.571 0 0 1 -.571.571h-2.286a.571.571 0 0 1 -.572-.571.571.571 0 0 1 .572-.571z"
-          opacity=".5"
-        />
-        <path
-          d="m25.143 8.571h2.286a.571.571 0 0 1 .571.572.571.571 0 0 1 -.571.571h-2.286a.571.571 0 0 1 -.571-.571.571.571 0 0 1 .571-.572z"
-          opacity=".5"
-        />
-        <path
-          d="m26.286 10.857h1.143a.571.571 0 0 1 .571.571.571.571 0 0 1 -.571.571h-1.143a.571.571 0 0 1 -.571-.571.571.571 0 0 1 .571-.571z"
-          opacity=".5"
-        />
-        <path
-          d="m26.286 13.143h1.143a.571.571 0 0 1 .571.571.571.571 0 0 1 -.571.571h-1.143a.571.571 0 0 1 -.571-.571.571.571 0 0 1 .571-.571z"
-          opacity=".5"
-        />
-        <path
-          d="m17.143 22.286h10.286a.571.571 0 0 1 .571.571.571.571 0 0 1 -.571.571h-10.286a.571.571 0 0 1 -.571-.571.571.571 0 0 1 .571-.571z"
-          opacity=".5"
-        />
-        <path
-          d="m12.571 22.286h2.286a.571.571 0 0 1 .571.571.571.571 0 0 1 -.571.571h-2.286a.571.571 0 0 1 -.571-.571.571.571 0 0 1 .571-.571z"
-          opacity=".5"
-        />
-        <path
-          d="m21.714 26.857h5.714a.571.571 0 0 1 .571.571.571.571 0 0 1 -.571.571h-5.714a.571.571 0 0 1 -.571-.571.571.571 0 0 1 .571-.571z"
-          opacity=".5"
-        />
-        <path
-          d="m12.571 26.857h6.857a.571.571 0 0 1 .571.571.571.571 0 0 1 -.571.571h-6.857a.571.571 0 0 1 -.571-.57.571.571 0 0 1 .571-.572z"
-          opacity=".5"
-        />
-        <path
-          d="m12.571 10.857h3.429a.571.571 0 0 1 .571.571.571.571 0 0 1 -.571.572h-3.429a.571.571 0 0 1 -.571-.571.571.571 0 0 1 .571-.572z"
-          opacity=".5"
-        />
-        <path
-          d="m12.571 13.143h3.429a.571.571 0 0 1 .571.571.571.571 0 0 1 -.571.571h-3.429a.571.571 0 0 1 -.571-.571.571.571 0 0 1 .571-.571z"
-          opacity=".5"
-        />
-      </g>
-    </svg>
-  ),
-
-  prettier_tailwind: (
-    <svg
-      className="h-6 w-6"
-      height="24"
-      viewBox="0 0 24 24"
-      width="24"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path d="m12 6c-2.67 0-4.33 1.33-5 4 1-1.33 2.17-1.83 3.5-1.5.76.19 1.31.74 1.91 1.35.98 1 2.09 2.15 4.59 2.15 2.67 0 4.33-1.33 5-4-1 1.33-2.17 1.83-3.5 1.5-.76-.19-1.3-.74-1.91-1.35-.98-1-2.09-2.15-4.59-2.15m-5 6c-2.67 0-4.33 1.33-5 4 1-1.33 2.17-1.83 3.5-1.5.76.19 1.3.74 1.91 1.35.98 1 2.09 2.15 4.59 2.15 2.67 0 4.33-1.33 5-4-1 1.33-2.17 1.83-3.5 1.5-.76-.19-1.3-.74-1.91-1.35-.98-1-2.09-2.15-4.59-2.15z" />
-    </svg>
-  ),
-
-  pretier_sort: (
-    <svg
-      className="h-6 w-6"
-      height="512"
-      width="439"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path d="m304.804657 24.3809528h-292.6596298c-16.5529213-.7516956-15.8297806-24.246706 0-24.3809528h292.6596298c15.9111938.2409473 16.4802246 23.565073 0 24.3809528zm-292.66329 97.5238095c-16.5479336-.7517014-15.8250103-24.2467041 0-24.3809509h170.7621918c15.9063873.2409515 16.4752655 23.5650711 0 24.3809509zm121.9047603-48.7619019c-16.5479355-.7517014-15.8250122-24.2467079 0-24.3809547l243.9050446.0000038c15.9064026.2409477 16.4752808 23.5650711 0 24.3809509zm195.1408539 24.3809509c15.9063721.2409515 16.4775696 23.5650711.0022888 24.3809509h-97.619339c-16.5479279-.7517014-15.8249969-24.2467041 0-24.3809509zm48.6686707 24.380951c-16.5479431-.7517014-15.8250122-24.2467041 0-24.3809509h24.4764709c15.9064026.2409515 16.4752808 23.5650711 0 24.3809509zm-365.7142849 48.7619095c-16.5479336-.751709-15.8250103-24.2467194 0-24.3809662l48.8574371.0000153c15.9063797.2409515 16.4752655 23.5650635 0 24.3809509zm292.5714321 0c-16.5479431-.751709-15.8250122-24.2467194 0-24.3809662l122.0002747.0000153c15.9064026.2409515 16.4752808 23.5650635 0 24.3809509zm-195.0476227 0c-16.5479355-.751709-15.8250122-24.2467194 0-24.3809662l24.4764862.0000153c15.9063721.2409515 16.4752655 23.5650635 0 24.3809509zm-97.5238094 292.5714263c-16.5479336-.751709-15.8250103-24.2467041 0-24.3809509h48.8574371c15.9063797.2409363 16.4752655 23.5650635 0 24.3809509zm97.5238094 0c-16.5479355-.751709-15.8250122-24.2467041 0-24.3809509h24.4764862c15.9063721.2409363 16.4752655 23.5650635 0 24.3809509zm195.0476227-243.8095245c-16.5479431-.7516937-15.8250122-24.2467041 0-24.3809509h122.0002747c15.9064026.2409515 16.4752808 23.5650635 0 24.3809509zm-292.5714321 0c-16.5479336-.7516937-15.8250103-24.2467041 0-24.3809509h122.0002975c15.9063721.2409515 16.4752655 23.5650635 0 24.3809509zm0 195.0476227c-16.5479336-.751709-15.8250103-24.2467041 0-24.3809509h122.0002975c15.9063721.2409363 16.4752655 23.5650635 0 24.3809509zm0 97.5238037c-16.5479336-.751709-15.8250103-24.2467041 0-24.3809509h122.0002975c15.9063721.2409363 16.4752655 23.5650635 0 24.3809509zm0-243.8095398c-16.5479336-.7516785-15.8250103-24.2466888 0-24.3809357h24.4764843c15.9063797.2409515 16.4752617 23.5650787 0 24.3809662zm73.1428585 0c-16.5479355-.7516785-15.8250122-24.2466888 0-24.3809357h73.2383881c15.9063873.2409515 16.4752655 23.5650787 0 24.3809662zm121.9047546 0c-16.5479279-.7516785-15.8250122-24.2466888 0-24.3809357h195.1431427c15.9064026.2409515 16.4752808 23.5650787 0 24.3809662zm97.523819 48.7619324c-16.5479431-.751709-15.8250122-24.2467041 0-24.3809509h73.2383728c15.9064026.2409363 16.4752808 23.5650635 0 24.3809509zm-146.2857209 0c-16.5479279-.751709-15.8250122-24.2467041 0-24.3809509h97.619339c15.9063721.2409363 16.4752808 23.5650635 0 24.3809509zm-146.2857112 0c-16.5479336-.751709-15.8250103-24.2467041 0-24.3809509h97.6193466c15.9063797.2409363 16.4752579 23.5650635 0 24.3809509zm0 48.7619018c-16.5479336-.751709-15.8250103-24.2467041 0-24.3809509h24.4764843c15.9063797.2409363 16.4752617 23.5650635 0 24.3809509zm73.1428585 0c-16.5479355-.751709-15.8250122-24.2467041 0-24.3809509h219.5240936c15.9064026.2409363 16.4752808 23.5650635 0 24.3809509zm-73.1428585-292.571434c-16.5479336-.7517014-15.8250103-24.2467079 0-24.3809547l73.1919708.0000038c15.967186.2409477 16.5382385 23.5650711 0 24.3809509z" />
-    </svg>
-  ),
-
-  husky: (
-    <svg
-      className="h-6 w-6"
-      fill="none"
-      height="400"
-      viewBox="0 0 400 400"
-      width="400"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <g
-        stroke="#000"
-        stroke-linecap="round"
-        stroke-linejoin="round"
-        stroke-opacity=".9"
-        stroke-width="16"
-      >
-        <path d="m117.635 180.734c-6.332-11.533-1.139-135.4387 26.676-119.2839 22.883 13.2893 21.582 55.9299 31.128 68.8639" />
-        <path d="m220.284 126.353c55.863-128.19337 73.607-4.96 66.375 56.95" />
-        <path d="m96 247.215c4.992-25.717 27.707-19.999 42.025-30.492 16.971-12.443 20.302-29.251 38.839-29.251 18.538 0 24.615 10.873 25.988 10.1 7.834-4.408 11.427-14.885 31.14-10.755 19.713 4.129 13.432 24.465 24.37 33.019 18.669 14.595 37.456 10.19 45.206 37.962" />
-        <path d="m99.8438 269.555c59.2972-5.16 53.9402 93.003 120.8982 65.61 28.87-11.813 34.635-49.878 67.45-57.711" />
-        <path d="m169.033 225.688c3.328 2.432 7.378 3.676 11.531 5.138" />
-        <path d="m220.284 230.826c5.061-.91 8.773-2.391 12.813-5.138" />
-        <path d="m202.521 287.415c32.04-9.631-7.382 36.197-3.784 3.213" />
-      </g>
-    </svg>
-  ),
-
-  facebook: (
-    <svg
-      className="h-5 w-5"
-      viewBox="0 0 512 512"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <rect fill="#1877f2" height="512" rx="15%" width="512" />
-      <path
-        d="m355.6 330 11.4-74h-71v-48c0-20.2 9.9-40 41.7-40h32.3v-63s-29.3-5-57.3-5c-58.5 0-96.7 35.4-96.7 99.6v56.4h-65v74h65v182h80v-182z"
-        fill="#fff"
-      />
-    </svg>
-  ),
-
-  linkedin: (
-    <svg
-      className="h-5 w-5"
-      viewBox="0 0 128 128"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="m116 3h-104a8.91 8.91 0 0 0 -9 8.8v104.42a8.91 8.91 0 0 0 9 8.78h104a8.93 8.93 0 0 0 9-8.81v-104.42a8.93 8.93 0 0 0 -9-8.77z"
-        fill="#0076b2"
-      />
-      <g fill="#fff">
-        <path d="m21.06 48.73h18.11v58.27h-18.11zm9.06-29a10.5 10.5 0 1 1 -10.5 10.49 10.5 10.5 0 0 1 10.5-10.49" />
-        <path d="m50.53 48.73h17.36v8h.24c2.42-4.58 8.32-9.41 17.13-9.41 18.34-.04 21.74 12.03 21.74 27.68v32h-18.11v-28.35c0-6.75-.12-15.44-9.41-15.44s-10.87 7.36-10.87 15v28.79h-18.08z" />
-      </g>
-    </svg>
-  ),
-
-  twitter: (
-    <svg
-      className="h-5 w-5"
-      viewBox="0 0 126.24 102.59"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="m40.58 115.3c47.64 0 73.69-39.47 73.69-73.69 0-1.12 0-2.24-.07-3.35a52.7 52.7 0 0 0 12.92-13.41 51.7 51.7 0 0 1 -14.87 4.08 26 26 0 0 0 11.38-14.33 51.9 51.9 0 0 1 -16.45 6.29 25.92 25.92 0 0 0 -44.13 23.62 73.53 73.53 0 0 1 -53.38-27.06 25.92 25.92 0 0 0 8 34.58 25.71 25.71 0 0 1 -11.67-3.25v.33a25.91 25.91 0 0 0 20.73 25.39 25.86 25.86 0 0 1 -11.7.44 25.93 25.93 0 0 0 24.2 18 52 52 0 0 1 -32.17 11.06 52.72 52.72 0 0 1 -6.18-.36 73.32 73.32 0 0 0 39.7 11.63"
-        fill="#1da1f2"
-        transform="translate(-.88 -12.7)"
-      />
-    </svg>
-  ),
-
-  github: (
-    <svg
-      className="h-5 w-5"
-      viewBox="0 0 24 24"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="m0 0h24v24h-24z"
-        fill="#fff"
-        opacity="0"
-        transform="matrix(-1 0 0 -1 24 24)"
-      />
-      <path
-        d="m12 1a10.89 10.89 0 0 0 -11 10.77 10.79 10.79 0 0 0 7.52 10.23c.55.1.75-.23.75-.52s0-.93 0-1.83c-3.06.65-3.71-1.44-3.71-1.44a2.86 2.86 0 0 0 -1.22-1.58c-1-.66.08-.65.08-.65a2.31 2.31 0 0 1 1.68 1.11 2.37 2.37 0 0 0 3.2.89 2.33 2.33 0 0 1 .7-1.44c-2.44-.27-5-1.19-5-5.32a4.15 4.15 0 0 1 1.11-2.91 3.78 3.78 0 0 1 .11-2.84s.93-.29 3 1.1a10.68 10.68 0 0 1 5.5 0c2.1-1.39 3-1.1 3-1.1a3.78 3.78 0 0 1 .11 2.84 4.15 4.15 0 0 1 1.17 2.89c0 4.14-2.58 5.05-5 5.32a2.5 2.5 0 0 1 .75 2v2.95s.2.63.75.52a10.8 10.8 0 0 0 7.5-10.22 10.89 10.89 0 0 0 -11-10.77"
-        fill="#231f20"
-      />
-    </svg>
-  ),
-};
 
 const TECH_STACK = [
   {
@@ -431,7 +49,6 @@ const TECH_STACK = [
     description:
       'Utilize the power of Next.js for server-rendered React applications.',
     link: 'https://nextjs.org/',
-    icon: ICONS.next_js,
   },
 
   {
@@ -440,25 +57,22 @@ const TECH_STACK = [
     description:
       "Enhance your development experience with TypeScript's static typing.",
     link: 'https://www.typescriptlang.org/',
-    icon: ICONS.typescript,
   },
 
   {
     id: 3,
-    name: 'TailWind CSS',
+    name: 'Tailwind CSS',
     description:
       "Rapidly build custom designs with Tailwind CSS's utility-first approach.",
     link: 'https://tailwindcss.com/',
-    icon: ICONS.tailwind,
   },
 
   {
     id: 4,
-    name: 'ES Lint',
+    name: 'ESLint',
     description:
       'Enforce code quality standards and catch errors early with ESLint.',
     link: 'https://eslint.org/',
-    icon: ICONS.eslint,
   },
 
   {
@@ -467,7 +81,6 @@ const TECH_STACK = [
     description:
       'Keep your codebase clean and consistent with automatic code formatting using Prettier.',
     link: 'https://prettier.io/',
-    icon: ICONS.prettier,
   },
 
   {
@@ -476,7 +89,6 @@ const TECH_STACK = [
     description:
       'Automatic sorting of tailwind classnames using the official prettier plugin.',
     link: 'https://github.com/tailwindlabs/prettier-plugin-tailwindcss',
-    icon: ICONS.tailwind,
   },
 
   {
@@ -485,7 +97,6 @@ const TECH_STACK = [
     description:
       'Sort imports alphabetically for better readability, especially in larger projects.',
     link: 'https://github.com/IanVS/prettier-plugin-sort-imports',
-    icon: ICONS.prettier,
   },
 
   {
@@ -494,36 +105,5 @@ const TECH_STACK = [
     description:
       'Ensure code quality and prevent bad commits with pre-commit hooks powered by Husky.',
     link: 'https://typicode.github.io/husky/',
-    icon: ICONS.husky,
-  },
-];
-
-const SOCIAL_ICONS = [
-  {
-    id: 1,
-    name: 'Facebook',
-    link: 'https://www.facebook.com/creowis',
-    icon: ICONS.facebook,
-  },
-
-  {
-    id: 2,
-    name: 'Twitter',
-    link: 'https://twitter.com/creowistech',
-    icon: ICONS.twitter,
-  },
-
-  {
-    id: 3,
-    name: 'LinkedIn',
-    link: 'https://www.linkedin.com/company/creowis/about/',
-    icon: ICONS.linkedin,
-  },
-
-  {
-    id: 4,
-    name: 'Github',
-    link: 'https://github.com/',
-    icon: ICONS.github,
   },
 ];
