+++ src/components/copy-button/index.ts
@@ -1,0 +1,1 @@
+export { CopyButton as default } from './copy-button';
